WIP - Move swapchain setup out of device
