+++ src/OpenH2.Rendering/Vulkan/VkDevice.cs
@@ -13,35 +13,34 @@ namespace OpenH2.Rendering.Vulkan
     {
         private Device device;
         private VkInstance instance;
-        private KhrSwapchain khrSwapchainExt;
         private KhrSurface khrSurfaceExtension;
 
         private SurfaceKHR surface;
+        private SurfaceFormatKHR format;
 
         private (uint? graphics, uint? present) queueFamilies;
         private Queue graphicsQueue;
         private Queue presentQueue;
-        private SwapchainKHR swapchain;
-        private Image[] swapchainImages;
-        private ImageView[] swapchainImageviews;
-        private Framebuffer[] swapchainFramebuffers;
+
         private (Format format, Extent2D extent) swapchainParams;
         private readonly VulkanHost vulkanHost;
 
         public KhrSurface KhrSurfaceExtension => khrSurfaceExtension;
         public SurfaceKHR Surface => surface;
-
-        public Extent2D Extent => swapchainParams.extent;
-        public Format Format => swapchainParams.format;
+        public SurfaceFormatKHR SurfaceFormat => format;
 
         public uint? GraphicsQueueFamily => queueFamilies.graphics;
         public Queue GraphicsQueue => graphicsQueue;
+        public uint? PresentQueueFamily => queueFamilies.present;
         public Queue PresentQueue => presentQueue;
 
-        public SwapchainKHR Swapchain => swapchain;
-        public Image[] Images => swapchainImages;
-        public ImageView[] ImageViews => swapchainImageviews;
-        public Framebuffer[] Framebuffers => swapchainFramebuffers;
+        private PresentModeKHR presentMode;
+        public PresentModeKHR PresentMode => presentMode;
+
+        private SurfaceCapabilitiesKHR caps;
+        public SurfaceCapabilitiesKHR SurfaceCapabilities => caps;
+
+        public Extent2D Extent => this.ChooseSwapExtent(this.caps);
 
         public VkDevice(VulkanHost host, VkInstance instance, string[] validationLayers) : base(instance.vk)
         {
@@ -52,7 +51,7 @@ namespace OpenH2.Rendering.Vulkan
             this.surface = host.window.VkSurface.Create<AllocationCallbacks>(instance.Instance.ToHandle(), null).ToSurface();
 
             // Choose appropriate physical device and get relevant settings to create logical device and swapchains
-            if (!ChooseDevice(instance, out var phyDevice, out queueFamilies, out var caps, out var format, out var presentMode, out var extent))
+            if (!ChooseDevice(instance, out var phyDevice, out queueFamilies, out caps, out format, out presentMode))
             {
                 throw new Exception("No supported devices found");
             }
@@ -99,86 +98,15 @@ namespace OpenH2.Rendering.Vulkan
 
             // Create logical device and grab device-specific resources
             SUCCESS(vk.CreateDevice(phyDevice, in deviceCreate, null, out device), "Logical device creation failed");
-            SUCCESS(vk.TryGetDeviceExtension(instance, device, out khrSwapchainExt), "Couldn't get swapchain extension");
             vk.GetDeviceQueue(device, queueFamilies.graphics.Value, 0, out graphicsQueue);
             vk.GetDeviceQueue(device, queueFamilies.present.Value, 0, out presentQueue);
-
-            // One more than min count, unless we're bound by max image count
-            var imgCount = caps.MinImageCount + 1;
-            if (caps.MaxImageCount > 0 && imgCount > caps.MaxImageCount)
-                imgCount = caps.MaxImageCount;
-
-            // Prepare to create swapchain with what we've found
-            var swapchainCreate = new SwapchainCreateInfoKHR
-            {
-                SType = StructureType.SwapchainCreateInfoKhr,
-                Surface = surface,
-                MinImageCount = imgCount,
-                ImageFormat = format.Format,
-                ImageColorSpace = format.ColorSpace,
-                ImageExtent = extent,
-                ImageArrayLayers = 1,
-                ImageUsage = ImageUsageFlags.ImageUsageColorAttachmentBit,
-                PreTransform = caps.CurrentTransform,
-                CompositeAlpha = CompositeAlphaFlagsKHR.CompositeAlphaOpaqueBitKhr,
-                PresentMode = presentMode,
-                Clipped = true,
-                OldSwapchain = default,
-                ImageSharingMode = SharingMode.Exclusive,
-            };
-
-            // If we're not using the same queue, setup concurrent images
-            if (queueFamilies.graphics != queueFamilies.present)
-            {
-                var queueFamilyIndices = stackalloc uint[] { queueFamilies.graphics.Value, queueFamilies.present.Value };
-                swapchainCreate.ImageSharingMode = SharingMode.Concurrent;
-                swapchainCreate.QueueFamilyIndexCount = 2;
-                swapchainCreate.PQueueFamilyIndices = queueFamilyIndices;
-            }
-
-            SUCCESS(khrSwapchainExt.CreateSwapchain(device, in swapchainCreate, null, out swapchain), "failed to create swapchain");
-            khrSwapchainExt.GetSwapchainImages(device, swapchain, ref imgCount, null);
-            swapchainImages = new Image[imgCount];
-            swapchainImageviews = new ImageView[imgCount];
-            swapchainFramebuffers = new Framebuffer[imgCount];
-
-            khrSwapchainExt.GetSwapchainImages(device, swapchain, ref imgCount, out swapchainImages[0]);
-            swapchainParams = (format.Format, extent);
-
-            for (int i = 0; i < imgCount; i++)
-            {
-                var imgViewCreate = new ImageViewCreateInfo
-                {
-                    SType = StructureType.ImageViewCreateInfo,
-                    Image = swapchainImages[i],
-                    ViewType = ImageViewType.ImageViewType2D,
-                    Format = format.Format,
-                    Components = new ComponentMapping(ComponentSwizzle.Identity, ComponentSwizzle.Identity, ComponentSwizzle.Identity, ComponentSwizzle.Identity),
-                    SubresourceRange = new ImageSubresourceRange
-                    {
-                        AspectMask = ImageAspectFlags.ImageAspectColorBit,
-                        BaseMipLevel = 0,
-                        LevelCount = 1,
-                        BaseArrayLayer = 0,
-                        LayerCount = 1
-                    }
-                };
-
-                SUCCESS(vk.CreateImageView(device, in imgViewCreate, null, out swapchainImageviews[i]), "Image view create failed");
-            }
         }
 
-        public Result AcquireNextImage(Semaphore semaphore, Fence fence, ref uint index)
+        public VkSwapchain CreateSwapchain()
         {
-            return khrSwapchainExt.AcquireNextImage(device, swapchain, ulong.MaxValue, semaphore, fence, ref index);
+            return new VkSwapchain(this.vulkanHost, this.instance, this);
         }
 
-        public Result QueuePresent(in PresentInfoKHR presentInfo)
-        {
-            return khrSwapchainExt.QueuePresent(presentQueue, in presentInfo);
-        }
-
-        // TODO: move to instance?
         private bool QuerySwapchainSupport(PhysicalDevice phyDevice,
             out SurfaceCapabilitiesKHR capabilities,
             out SurfaceFormatKHR[] formats,
@@ -239,14 +167,12 @@ namespace OpenH2.Rendering.Vulkan
             out (uint? graphics, uint? present) queueFamilies,
             out SurfaceCapabilitiesKHR capabilities,
             out SurfaceFormatKHR format,
-            out PresentModeKHR presentMode,
-            out Extent2D extent)
+            out PresentModeKHR presentMode)
         {
             queueFamilies = default;
             capabilities = default;
             format = default;
             presentMode = PresentModeKHR.PresentModeFifoKhr;
-            extent = default;
 
             uint deviceCount = 0;
             vk.EnumeratePhysicalDevices(instance, ref deviceCount, null);
@@ -273,7 +199,6 @@ namespace OpenH2.Rendering.Vulkan
 
                 format = ChooseFormat(formats);
                 presentMode = ChoosePresentMode(presentModes);
-                extent = ChooseSwapExtent(capabilities);
 
                 if (!queueFamilies.graphics.HasValue)
                     continue;
@@ -326,17 +251,6 @@ namespace OpenH2.Rendering.Vulkan
 
         public void Dispose()
         {
-            foreach (var imgView in swapchainImageviews)
-            {
-                vk.DestroyImageView(this.device, imgView, null);
-            }
-
-            if (khrSwapchainExt != null)
-            {
-                khrSwapchainExt.DestroySwapchain(device, swapchain, null);
-                khrSwapchainExt.Dispose();
-            }
-
             if (khrSurfaceExtension != null)
             {
                 khrSurfaceExtension.DestroySurface(instance, surface, null);
@@ -345,7 +259,5 @@ namespace OpenH2.Rendering.Vulkan
 
             vk.DestroyDevice(this.device, null);
         }
-
-        
     }
 }
